+++ src/main/java/kitchenpos/menugroup/domain/MenuGroupRepository.java
@@ -1,6 +1,5 @@
-package kitchenpos.menugroup.repository;
+package kitchenpos.menugroup.domain;
 
-import kitchenpos.menugroup.domain.MenuGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MenuGroupRepository extends JpaRepository<MenuGroup, Long> {
