+++ src/main/java/kitchenpos/table/domain/OrderTableRepository.java
@@ -1,6 +1,5 @@
-package kitchenpos.table.repository;
+package kitchenpos.table.domain;
 
-import kitchenpos.table.domain.OrderTable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
