+++ src/main/java/kitchenpos/tablegroup/domain/TableGroupRepository.java
@@ -1,6 +1,5 @@
-package kitchenpos.tablegroup.repository;
+package kitchenpos.tablegroup.domain;
 
-import kitchenpos.tablegroup.domain.TableGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface TableGroupRepository extends JpaRepository<TableGroup, Long> {
