+++ src/test/java/kitchenpos/menu/application/MenuServiceTest.java
@@ -4,9 +4,9 @@ import kitchenpos.menu.domain.Menu;
 import kitchenpos.menugroup.domain.MenuGroup;
 import kitchenpos.menu.domain.MenuProduct;
 import kitchenpos.product.domain.Product;
-import kitchenpos.menugroup.repository.MenuGroupRepository;
-import kitchenpos.menu.repository.MenuRepository;
-import kitchenpos.product.repository.ProductRepository;
+import kitchenpos.menugroup.domain.MenuGroupRepository;
+import kitchenpos.menu.domain.MenuRepository;
+import kitchenpos.product.domain.ProductRepository;
 import kitchenpos.menu.dto.CreateMenuProductRequest;
 import kitchenpos.menu.dto.CreateMenuRequest;
 import org.assertj.core.api.SoftAssertions;
