+++ src/test/java/kitchenpos/table/application/TableServiceTest.java
@@ -10,12 +10,12 @@ import kitchenpos.table.domain.NumberOfGuests;
 import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.table.domain.OrderTable;
 import kitchenpos.tablegroup.domain.TableGroup;
-import kitchenpos.menugroup.repository.MenuGroupRepository;
-import kitchenpos.menu.repository.MenuRepository;
-import kitchenpos.order.repository.OrderRepository;
-import kitchenpos.table.repository.OrderTableRepository;
-import kitchenpos.product.repository.ProductRepository;
-import kitchenpos.tablegroup.repository.TableGroupRepository;
+import kitchenpos.menugroup.domain.MenuGroupRepository;
+import kitchenpos.menu.domain.MenuRepository;
+import kitchenpos.order.domain.OrderRepository;
+import kitchenpos.table.domain.OrderTableRepository;
+import kitchenpos.product.domain.ProductRepository;
+import kitchenpos.tablegroup.domain.TableGroupRepository;
 import kitchenpos.table.dto.ChangeNumberOfGuestsRequest;
 import kitchenpos.table.dto.ChangeOrderTableEmptyRequest;
 import kitchenpos.table.dto.CreateOrderTableRequest;
