+++ src/test/java/kitchenpos/tablegroup/application/TableGroupServiceTest.java
@@ -9,12 +9,12 @@ import kitchenpos.order.domain.OrderLineItems;
 import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.table.domain.OrderTable;
 import kitchenpos.tablegroup.domain.TableGroup;
-import kitchenpos.menugroup.repository.MenuGroupRepository;
-import kitchenpos.menu.repository.MenuRepository;
-import kitchenpos.order.repository.OrderRepository;
-import kitchenpos.table.repository.OrderTableRepository;
-import kitchenpos.product.repository.ProductRepository;
-import kitchenpos.tablegroup.repository.TableGroupRepository;
+import kitchenpos.menugroup.domain.MenuGroupRepository;
+import kitchenpos.menu.domain.MenuRepository;
+import kitchenpos.order.domain.OrderRepository;
+import kitchenpos.table.domain.OrderTableRepository;
+import kitchenpos.product.domain.ProductRepository;
+import kitchenpos.tablegroup.domain.TableGroupRepository;
 import kitchenpos.tablegroup.dto.CreateTableGroupOrderTableRequest;
 import kitchenpos.tablegroup.dto.CreateTableGroupRequest;
 import org.assertj.core.api.SoftAssertions;
